refactor(web): split _shared.js god-file into focused modules
_shared.js was a ~790-line module mixing four unrelated concerns: utility
helpers, the SSRF guard, the entire RESULTS KV data layer, and HTML/SVG
presentation. Split by concern so each is independently readable and the data
layer is framework-agnostic (ready for the upcoming Hono routes to import):

  _util.js    ids, domain normalization, email check, tier ranking (leaf)
  _ssrf.js    scan-URL SSRF guard (private/loopback/internal blocking)
  _data.js    RESULTS KV: results, watches, tokens, history, stats, listings
  _render.js  tier colors, badge SVG, share-card HTML, HTML/XML escaping

_shared.js is now a re-export barrel, so all 28 existing call sites keep their
`from '../_shared.js'` imports unchanged. Pure mechanical move — no logic edits;
the full suite (198 tests) passes identically.
